start refactoring (review)
